embedding/pinecone/requirements.txt
# Pinecone migration requirements
# Core dependencies from parent requirements.txt

# Pinecone client and dependencies
pinecone-client>=3.0.0
numpy>=1.21.0
tqdm>=4.64.0

# Additional dependencies for migration
python-dotenv>=0.19.0
